Basic Uplink Ingest

+++ priv/repo/migrations/20210117225810_create_uplinks_heard.exs
@@ -8,7 +8,7 @@ defmodule Mappers.Repo.Migrations.CreateUplinksHeard do
       add :hotspot_name, :string
       add :latitude, :float
       add :longitude, :float
-      add :rssi, :integer
+      add :rssi, :float
       add :snr, :float
       add :timestamp, :utc_datetime_usec
       add :uplink_id, references(:uplinks, type: :uuid, on_delete: :delete_all)
